systhreads: major refactor, removing mentions of DECs, streamlining the patch to reduce the diff to systhreads
