+++ otherlibs/systhreads/mutex.ml
@@ -1,5 +1,19 @@
-type t
+(**************************************************************************)
+(*                                                                        *)
+(*                                 OCaml                                  *)
+(*                                                                        *)
+(*           Xavier Leroy and Pascal Cuoq, INRIA Rocquencourt             *)
+(*                                                                        *)
+(*   Copyright 1995 Institut National de Recherche en Informatique et     *)
+(*     en Automatique.                                                    *)
+(*                                                                        *)
+(*   All rights reserved.  This file is distributed under the terms of    *)
+(*   the GNU Lesser General Public License version 2.1, with the          *)
+(*   special exception on linking described in the file LICENSE.          *)
+(*                                                                        *)
+(**************************************************************************)
 
+type t
 external create: unit -> t = "caml_mutex_new"
 external lock: t -> unit = "caml_mutex_lock"
 external try_lock: t -> bool = "caml_mutex_try_lock"
